requirements/base.txt
Django==1.8
psycopg2==2.6.1
py==1.4.31
pyTelegramBotAPI==1.4.1
pytest==2.8.5
requests==2.9.1
six==1.10.0
wheel==0.24.0
django-grappelli==2.7.3
PyYaml==3.11